requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
langchain==0.0.335
langchain-community==0.0.13
python-dotenv==1.0.0
pydantic==2.4.2
openai==1.3.0 